+++ refinedstorage-common/src/main/java/com/refinedmods/refinedstorage/common/autocrafting/monitor/AutocraftingTaskButton.java
@@ -61,13 +61,13 @@ class AutocraftingTaskButton extends AbstractButton {
     protected void renderWidget(final GuiGraphics graphics,
                                 final int mouseX,
                                 final int mouseY,
-                                final float partialTick) {
-        super.renderWidget(graphics, mouseX, mouseY, partialTick);
+                                final float partialTicks) {
+        super.renderWidget(graphics, mouseX, mouseY, partialTicks);
         renderResourceIcon(graphics);
         final int yOffset = SmallText.isSmall() ? 5 : 3;
         final int textX = getX() + 3 + 16 + 3;
         final int textY = getY() + yOffset;
-        text.render(graphics, textX, textY, Minecraft.getInstance().font, isHovered);
+        text.render(graphics, textX, textY, Minecraft.getInstance().font, isHovered, partialTicks);
         final int ySpacing = SmallText.isSmall() ? 7 : 8;
         final long percentageCompleted = Math.round(
             stateProvider.getPercentageCompleted(task.id()) * 100